requirements.txt
requests
python-dotenv
beautifulsoup4
pdfplumber
python-docx